+++ Directx_11_Game_Engine/RenderWindow.h
@@ -6,7 +6,7 @@ class WindowContainer;
 class RenderWindow
 {
 public:
-	bool Initialise(WindowContainer* pWindowContainer, HINSTANCE hInstance, std::string window_title, std::string window_class, float width, float height);
+	bool Initialise(WindowContainer* pWindowContainer, HINSTANCE hInstance, std::string window_title, std::string window_class, int width, int height);
 	bool ProcessMessages();
 	HWND GetHWND() const;
 	~RenderWindow();
@@ -20,8 +20,8 @@ private:
 	std::wstring window_title_wide = L""; //wide string for windows title
 	std::string window_class = "";
 	std::wstring window_class_wide = L""; //wide string for window class
-	float width = 0.0f;
-	float height = 0.0f;
+	int width = 0;
+	int height = 0;
 
 };
 LRESULT CALLBACK HandleMsgRedirect(HWND hwnd, UINT uMsg, WPARAM wParam, LPARAM lParam);
